+++ lib/kitchen/directions/bake_example.rb
@@ -3,22 +3,21 @@
 module Kitchen
   module Directions
     module BakeExample
-      def self.v1(example:, number:, title_tag:, numbered_solutions: false)
+      def self.v1(example:, number:, title_tag:, numbered_solutions: false, cases: false)
         example.wrap_children(class: 'body')
 
         example.prepend(child:
           <<~HTML
             <#{title_tag} class="os-title">
-              <span class="os-title-label">#{I18n.t(:example_label)} </span>
+              <span class="os-title-label">#{I18n.t("example#{'.nominative' if cases}")} </span>
               <span class="os-number">#{number}</span>
               <span class="os-divider"> </span>
             </#{title_tag}>
           HTML
         )
 
-        example.document
-               .pantry(name: :link_text)
-               .store("#{I18n.t(:example_label)} #{number}", label: example.id)
+        # Store label information
+        example.target_label(label_text: 'example', custom_content: number, cases: cases)
 
         example.titles_to_rename.each do |title|
           title.name = 'h4'