+++ lib/kitchen/directions/bake_figure.rb
@@ -3,7 +3,7 @@
 module Kitchen
   module Directions
     module BakeFigure
-      def self.v1(figure:, number:)
+      def self.v1(figure:, number:, cases: false)
         return if figure.has_class?('unnumbered') && !figure.has_class?('splash')
 
         figure.wrap(%(<div class="os-figure#{' has-splash' if figure.has_class?('splash')}">))
@@ -19,14 +19,16 @@ module Kitchen
           return
         end
 
-        figure.pantry(name: :link_text).store "#{I18n.t(:figure)} #{number}", label: figure.id
+        # Store label information
+        figure.target_label(label_text: 'figure', custom_content: number, cases: cases)
+
         title = figure.title&.cut
 
         caption = figure.caption&.cut
         figure.append(sibling:
           <<~HTML
             <div class="os-caption-container">
-              <span class="os-title-label">#{I18n.t(:figure)} </span>
+              <span class="os-title-label">#{I18n.t("figure#{'.nominative' if cases}")} </span>
               <span class="os-number">#{number}</span>
               <span class="os-divider"> </span>
               #{"<span class=\"os-title\" data-type=\"title\" id=\"#{title.id}\">#{title.children}</span>" if title}