+++ lib/kitchen/directions/bake_notes/bake_note_subtitle.rb
@@ -3,14 +3,14 @@
 module Kitchen
   module Directions
     module BakeNoteSubtitle
-      def self.v1(note:)
+      def self.v1(note:, cases: false)
         title = note.title&.cut
 
         return unless title
 
         # Store label information
         note_label = title.children
-        note.pantry(name: :link_text).store note_label, label: note.id
+        note.target_label(custom_content: note_label.to_s, cases: cases)
 
         title.name = 'h4'
         title.add_class('os-subtitle')