+++ lib/kitchen/directions/bake_numbered_exercise/main.rb
@@ -3,9 +3,10 @@
 module Kitchen
   module Directions
     module BakeNumberedExercise
-      def self.v1(exercise:, number:, suppress_solution_if: false, note_suppressed_solutions: false)
+      def self.v1(exercise:, number:, suppress_solution_if: false,
+                  note_suppressed_solutions: false, cases: false)
         V1.new.bake(exercise: exercise, number: number, suppress_solution_if: suppress_solution_if,
-                    note_suppressed_solutions: note_suppressed_solutions)
+                    note_suppressed_solutions: note_suppressed_solutions, cases: cases)
       end
 
       def self.bake_solution_v1(exercise:, number:, divider: '. ')