+++ lib/kitchen/directions/bake_numbered_table/main.rb
@@ -3,12 +3,12 @@
 module Kitchen
   module Directions
     module BakeNumberedTable
-      def self.v1(table:, number:, always_caption: false)
-        V1.new.bake(table: table, number: number, always_caption: always_caption)
+      def self.v1(table:, number:, always_caption: false, cases: false)
+        V1.new.bake(table: table, number: number, always_caption: always_caption, cases: cases)
       end
 
-      def self.v2(table:, number:)
-        V2.new.bake(table: table, number: number)
+      def self.v2(table:, number:, cases: false)
+        V2.new.bake(table: table, number: number, cases: cases)
       end
     end
   end