+++ spec/directions/bake_link_placeholders_spec.rb
@@ -21,20 +21,52 @@ RSpec.describe Kitchen::Directions::BakeLinkPlaceholders do
     )
   end
 
-  before do
-    book.pantry(name: :link_text).store('Example x.y', label: 'key')
-  end
-
-  it 'bakes' do
-    described_class.v1(book: book)
-    expect(book.body).to match_normalized_html(
+  let(:book_with_cases) do
+    book_containing(html:
       <<~HTML
-        <body>
-          <a>skip this link</a>
-          <a href='?key'>Example x.y</a>
-        </body>
+        <a>skip this link</a>
+        <a href='?other_key'>[link]</a>
+        <a xmlns:cmlnle="http://katalysteducation.org/cmlnle/1.0" cmlnle:case="genitive" href='?other_key'>[link]</a>
       HTML
     )
+  end
+
+  context 'when book do not use cases' do
+    before do
+      book.pantry(name: :link_text).store('Example x.y', label: 'key')
+    end
+
+    it 'bakes' do
+      described_class.v1(book: book)
+      expect(book.body).to match_normalized_html(
+        <<~HTML
+          <body>
+            <a>skip this link</a>
+            <a href='?key'>Example x.y</a>
+          </body>
+        HTML
+      )
+    end
+  end
+
+  context 'when book uses cases' do
+    before do
+      book_with_cases.pantry(name: :nominative_link_text).store('Przykład x.y', label: 'other_key')
+      book_with_cases.pantry(name: :genitive_link_text).store('Przykładu x.y', label: 'other_key')
+    end
+
+    it 'bakes' do
+      described_class.v1(book: book_with_cases, cases: true)
+      expect(book_with_cases.body).to match_normalized_html(
+        <<~HTML
+          <body>
+            <a>skip this link</a>
+            <a href='?other_key'>Przykład x.y</a>
+            <a xmlns:cmlnle="http://katalysteducation.org/cmlnle/1.0" cmlnle:case="genitive" href='?other_key'>Przykładu x.y</a>
+          </body>
+        HTML
+      )
+    end
   end
 
   it 'logs a warning' do